Expiry Notification
This script checks whether the Active Directory module is installed and whether the given user has an e-mail assigned, if not, it will inform about it in the console. Then it checks how many days are left until the password expires. If less than 20, it will send a notification to the e-mail assigned in Active Directory.